Fixed & Improved grammatical accuracy of es.lua
The current version seems to have been made using the rough translations of the English counter-parts. This version was written by a native speaker, some texts might not be the same but that's to improve user experience and also to help distinguish which commands are for which groups. 

Some of the Item names have been tweaked to also help distinguish them from the rest. 

Subsequent revisions/improvements will be provided once it has been tested further, this version should work perfectly fine but might need some tweaking depending of each server's UI setups (Due to proper translated name lengths).